+++ CHANGELOG.md
@@ -24,3 +24,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
       - [`esbuild@0.21.5`](https://npmjs.com/package/esbuild/v/0.21.5)
       - [`tsup@8.1.0`](https://npmjs.com/package/tsup/v/8.1.0)
       - [`typescript@5.5.2`](https://npmjs.com/package/typescript/v/5.5.2)
+- Added [ESLint import/export syntax](https://npmjs.com/package/eslint-plugin-import), in PR [#13](https://github.com/compulim/react-scrolling-background/pull/13)
+- Added [`publint`](https://npmjs.com/package/publint), in PR [#13](https://github.com/compulim/react-scrolling-background/pull/13)